# contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: CAE Material Platform API
  summary: Material data management, immutable traceability, and CAE workflow platform.
  version: 0.27.0
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /api/v1/shear-relaxation-datasets:
    post:
      operationId: importReferenceShearRelaxationDataset
      tags: [datasets]
      summary: Preserve raw and normalized shear-relaxation Dataset revisions.
      description: Explicit user-confirmed time and shear-modulus columns and units are required.
      security: [{BearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "../datasets/reference-shear-relaxation-resources.schema.json#/$defs/ImportRequest"}
      responses:
        "201":
          description: Raw revision and normalized SI Parquet revision created.
          headers:
            ETag: {$ref: "#/components/headers/RevisionETag"}
          content:
            application/json:
              schema: {$ref: "../datasets/reference-shear-relaxation-resources.schema.json#/$defs/Response"}
        "401": {description: Authentication required.}
        "403": {description: Dataset write is not authorized.}
        "404": {description: Concrete Test Run or raw Artifact is not visible.}
        "409": {description: Immutable source or Dataset state conflicts.}
        "422": {description: Mapping, units, or relaxation curve values are invalid.}
        "503": {description: Dataset service unavailable.}
  /api/v1/material-states/{material_state_id}/shear-relaxation-datasets:
    get:
      operationId: listMaterialStateShearRelaxationDatasets
      tags: [datasets]
      summary: List current shear-relaxation Dataset heads for one Material State.
      security: [{BearerAuth: []}]
      parameters: [{$ref: "#/components/parameters/MaterialStateId"}]
      responses:
        "200":
          description: Tenant-scoped current Dataset heads.
          content:
            application/json:
              schema: {$ref: "../datasets/reference-shear-relaxation-resources.schema.json#/$defs/ListResponse"}
        "401": {description: Authentication required.}
        "403": {description: Dataset read is not authorized.}
        "503": {description: Dataset service unavailable.}
  /api/v1/shear-relaxation-datasets/{dataset_id}/preview:
    get:
      operationId: previewShearRelaxationDataset
      tags: [datasets]
      summary: Preview a bounded deterministic sample of the current immutable curve.
      security: [{BearerAuth: []}]
      parameters:
        - name: dataset_id
          in: path
          required: true
          schema: {type: string, format: uuid}
        - name: maximum_points
          in: query
          required: false
          schema: {type: integer, minimum: 3, maximum: 10000, default: 500}
      responses:
        "200":
          description: Raw original-unit or normalized SI curve preview.
          content:
            application/json:
              schema: {$ref: "../datasets/reference-shear-relaxation-resources.schema.json#/$defs/Preview"}
        "401": {description: Authentication required.}
        "403": {description: Dataset read is not authorized.}
        "404": {description: No visible Dataset matches the identifier.}
        "503": {description: Dataset service unavailable.}
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT access token
      description: RFC 6750 bearer access token; ID tokens are rejected.
  headers:
    RevisionETag:
      description: Strong ETag for one concrete immutable revision.
      required: true
      schema:
        type: string
        pattern: '^"revision:[1-9][0-9]*:sha256:[0-9a-f]{64}"$'
  parameters:
...
    MaterialStateId:
      name: material_state_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...